complete all test cases

+++ banking-system-main/src/main/java/com/javahongkong/bootcamp/exercise/AccountInterface.java
@@ -37,5 +37,5 @@ public interface AccountInterface {
 	/**
 	 * {@link Account#pin}.
 	 */
-	double getPin();
+	int getPin();
 }
